refactor parallelize
