+++ halo2_proofs/src/poly/commitment/msm.rs
@@ -1,5 +1,6 @@
 use super::Params;
-use crate::arithmetic::{best_multiexp, parallelize, CurveAffine};
+use crate::arithmetic::{best_multiexp, CurveAffine};
+use crate::multicore::parallelize;
 use ff::Field;
 use group::Group;
