+++ halo2_proofs/src/poly/commitment/prover.rs
@@ -4,8 +4,9 @@ use rand_core::RngCore;
 use super::super::{Coeff, Polynomial};
 use super::{Blind, Params};
 use crate::arithmetic::{
-    best_multiexp, compute_inner_product, eval_polynomial, parallelize, CurveAffine, FieldExt,
+    best_multiexp, compute_inner_product, eval_polynomial, CurveAffine, FieldExt,
 };
+use crate::multicore::parallelize;
 use crate::transcript::{EncodedChallenge, TranscriptWrite};
 
 use group::Curve;
